Refactor internal methods into relevant submodule
