+++ polycommit/src/lib.rs
@@ -41,7 +41,7 @@ use snarkvm_utilities::{
     serialize::*,
 };
 
-use core::{fmt::Debug, iter::FromIterator};
+use core::fmt::Debug;
 use rand_core::RngCore;
 
 #[cfg(not(feature = "std"))]
@@ -217,20 +217,6 @@ pub trait PolynomialCommitment<F: Field>: Sized + Clone + Debug {
         Self::Randomness: 'a,
         Self::Commitment: 'a;
 
-    /// Open with individual challenges
-    fn open_individual_opening_challenges<'a>(
-        ck: &Self::CommitterKey,
-        labeled_polynomials: impl IntoIterator<Item = &'a LabeledPolynomial<F>>,
-        commitments: impl IntoIterator<Item = &'a LabeledCommitment<Self::Commitment>>,
-        point: &'a F,
-        opening_challenges: &dyn Fn(u64) -> F,
-        rands: impl IntoIterator<Item = &'a Self::Randomness>,
-        rng: Option<&mut dyn RngCore>,
-    ) -> Result<Self::Proof, Self::Error>
-    where
-        Self::Randomness: 'a,
-        Self::Commitment: 'a;
-
     /// On input a list of labeled polynomials and a query set, `open` outputs a proof of evaluation
     /// of the polynomials at the points in the query set.
     fn batch_open<'a>(
@@ -290,77 +276,6 @@ pub trait PolynomialCommitment<F: Field>: Sized + Clone + Debug {
                 query_comms,
                 *query,
                 opening_challenge,
-                query_rands,
-                Some(rng),
-            )?;
-
-            end_timer!(proof_time);
-
-            proofs.push(proof);
-        }
-        end_timer!(open_time);
-
-        Ok(proofs.into())
-    }
-
-    /// Batch open with individual challenges.
-    fn batch_open_individual_opening_challenges<'a>(
-        ck: &Self::CommitterKey,
-        labeled_polynomials: impl IntoIterator<Item = &'a LabeledPolynomial<F>>,
-        commitments: impl IntoIterator<Item = &'a LabeledCommitment<Self::Commitment>>,
-        query_set: &QuerySet<F>,
-        opening_challenges: &dyn Fn(u64) -> F,
-        rands: impl IntoIterator<Item = &'a Self::Randomness>,
-        rng: Option<&mut dyn RngCore>,
-    ) -> Result<Self::BatchProof, Self::Error>
-    where
-        Self::Randomness: 'a,
-        Self::Commitment: 'a,
-    {
-        let rng = &mut crate::optional_rng::OptionalRng(rng);
-        let poly_rand_comm: BTreeMap<_, _> = labeled_polynomials
-            .into_iter()
-            .zip(rands)
-            .zip(commitments.into_iter())
-            .map(|((poly, r), comm)| (poly.label(), (poly, r, comm)))
-            .collect();
-
-        let open_time = start_timer!(|| format!(
-            "Opening {} polynomials at query set of size {}",
-            poly_rand_comm.len(),
-            query_set.len(),
-        ));
-
-        let mut query_to_labels_map = BTreeMap::new();
-
-        for (label, point) in query_set.iter() {
-            let labels = query_to_labels_map.entry(point).or_insert_with(BTreeSet::new);
-            labels.insert(label);
-        }
-
-        let mut proofs = Vec::new();
-        for (query, labels) in query_to_labels_map.into_iter() {
-            let mut query_polys: Vec<&'a LabeledPolynomial<_>> = Vec::new();
-            let mut query_rands: Vec<&'a Self::Randomness> = Vec::new();
-            let mut query_comms: Vec<&'a LabeledCommitment<Self::Commitment>> = Vec::new();
-
-            for label in labels {
-                let (polynomial, rand, comm) = poly_rand_comm.get(label).ok_or(Error::MissingPolynomial {
-                    label: label.to_string(),
-                })?;
-
-                query_polys.push(polynomial);
-                query_rands.push(rand);
-                query_comms.push(comm);
-            }
-
-            let proof_time = start_timer!(|| "Creating proof");
-            let proof = Self::open_individual_opening_challenges(
-                ck,
-                query_polys,
-                query_comms,
-                &query,
-                opening_challenges,
                 query_rands,
                 Some(rng),
             )?;
@@ -477,40 +392,6 @@ pub trait PolynomialCommitment<F: Field>: Sized + Clone + Debug {
         })
     }
 
-    /// Open combinations with individual challenges.
-    fn open_combinations_individual_opening_challenges<'a>(
-        ck: &Self::CommitterKey,
-        linear_combinations: impl IntoIterator<Item = &'a LinearCombination<F>>,
-        polynomials: impl IntoIterator<Item = &'a LabeledPolynomial<F>>,
-        commitments: impl IntoIterator<Item = &'a LabeledCommitment<Self::Commitment>>,
-        query_set: &QuerySet<F>,
-        opening_challenges: &dyn Fn(u64) -> F,
-        rands: impl IntoIterator<Item = &'a Self::Randomness>,
-        rng: Option<&mut dyn RngCore>,
-    ) -> Result<BatchLCProof<F, Self>, Self::Error>
-    where
-        Self::Randomness: 'a,
-        Self::Commitment: 'a,
-    {
-        let linear_combinations: Vec<_> = linear_combinations.into_iter().collect();
-        let polynomials: Vec<_> = polynomials.into_iter().collect();
-        let poly_query_set = lc_query_set_to_poly_query_set(linear_combinations.iter().copied(), query_set);
-        let poly_evals = evaluate_query_set(polynomials.iter().copied(), &poly_query_set);
-        let proof = Self::batch_open_individual_opening_challenges(
-            ck,
-            polynomials,
-            commitments,
-            &poly_query_set,
-            opening_challenges,
-            rands,
-            rng,
-        )?;
-        Ok(BatchLCProof {
-            proof,
-            evaluations: Some(poly_evals.values().copied().collect()),
-        })
-    }
-
     /// Checks that `evaluations` are the true evaluations at `query_set` of the
     /// linear combinations of polynomials committed in `commitments`.
     #[allow(clippy::too_many_arguments)]
@@ -581,78 +462,22 @@ pub trait PolynomialCommitment<F: Field>: Sized + Clone + Debug {
         Ok(true)
     }
 
-    /// check but with individual challenges
-    fn check_individual_opening_challenges<'a>(
-        vk: &Self::VerifierKey,
-        commitments: impl IntoIterator<Item = &'a LabeledCommitment<Self::Commitment>>,
-        point: &'a F,
-        values: impl IntoIterator<Item = F>,
-        proof: &Self::Proof,
-        opening_challenges: &dyn Fn(u64) -> F,
-        rng: Option<&mut dyn RngCore>,
-    ) -> Result<bool, Self::Error>
-    where
-        Self::Commitment: 'a;
-
-    /// batch_check but with individual challenges
-    fn batch_check_individual_opening_challenges<'a, R: RngCore>(
-        vk: &Self::VerifierKey,
+    /// On input a list of polynomials, linear combinations of those polynomials,
+    /// and a query set, `open_combination` outputs a proof of evaluation of
+    /// the combinations at the points in the query set.
+    fn open_combinations_individual_opening_challenges<'a>(
+        ck: &Self::CommitterKey,
+        linear_combinations: impl IntoIterator<Item = &'a LinearCombination<F>>,
+        polynomials: impl IntoIterator<Item = &'a LabeledPolynomial<F>>,
         commitments: impl IntoIterator<Item = &'a LabeledCommitment<Self::Commitment>>,
         query_set: &QuerySet<F>,
-        evaluations: &Evaluations<F>,
-        proof: &Self::BatchProof,
         opening_challenges: &dyn Fn(u64) -> F,
-        rng: &mut R,
-    ) -> Result<bool, Self::Error>
+        rands: impl IntoIterator<Item = &'a Self::Randomness>,
+        rng: Option<&mut dyn RngCore>,
+    ) -> Result<BatchLCProof<F, Self>, Self::Error>
     where
-        Self::Commitment: 'a,
-    {
-        let commitments: BTreeMap<_, _> = commitments.into_iter().map(|c| (c.label(), c)).collect();
-
-        let mut query_to_labels_map = BTreeMap::new();
-        for (label, point) in query_set.iter() {
-            let labels = query_to_labels_map.entry(point).or_insert_with(BTreeSet::new);
-            labels.insert(label);
-        }
-
-        // Implicit assumption: proofs are order in same manner as queries in
-        // `query_to_labels_map`.
-        let proofs: Vec<_> = proof.clone().into();
-        assert_eq!(proofs.len(), query_to_labels_map.len());
-
-        let mut result = true;
-        for ((point, labels), proof) in query_to_labels_map.into_iter().zip(proofs) {
-            let mut comms: Vec<&'_ LabeledCommitment<_>> = Vec::new();
-            let mut values = Vec::new();
-            for label in labels {
-                let commitment = commitments.get(label).ok_or(Error::MissingPolynomial {
-                    label: label.to_string(),
-                })?;
-
-                let v_i = evaluations
-                    .get(&(label.clone(), point.clone()))
-                    .ok_or(Error::MissingEvaluation {
-                        label: label.to_string(),
-                    })?;
-
-                comms.push(commitment);
-                values.push(*v_i);
-            }
-
-            let proof_time = start_timer!(|| "Checking per-query proof");
-            result &= Self::check_individual_opening_challenges(
-                vk,
-                comms,
-                &point,
-                values,
-                &proof,
-                opening_challenges,
-                Some(rng),
-            )?;
-            end_timer!(proof_time);
-        }
-        Ok(result)
-    }
+        Self::Randomness: 'a,
+        Self::Commitment: 'a;
 
     /// Check combinations with individual challenges.
     fn check_combinations_individual_opening_challenges<'a, R: RngCore>(
@@ -666,63 +491,7 @@ pub trait PolynomialCommitment<F: Field>: Sized + Clone + Debug {
         rng: &mut R,
     ) -> Result<bool, Self::Error>
     where
-        Self::Commitment: 'a,
-    {
-        let BatchLCProof { proof, evaluations } = proof;
-
-        let lc_s = BTreeMap::from_iter(linear_combinations.into_iter().map(|lc| (lc.label(), lc)));
-
-        let poly_query_set = lc_query_set_to_poly_query_set(lc_s.values().copied(), eqn_query_set);
-        let poly_evals: Evaluations<_> = poly_query_set
-            .iter()
-            .cloned()
-            .zip(evaluations.clone().unwrap())
-            .collect();
-
-        for &(ref lc_label, ref point) in eqn_query_set {
-            if let Some(lc) = lc_s.get(lc_label) {
-                let claimed_rhs =
-                    *eqn_evaluations
-                        .get(&(lc_label.clone(), point.clone()))
-                        .ok_or(Error::MissingEvaluation {
-                            label: lc_label.to_string(),
-                        })?;
-
-                let mut actual_rhs = F::zero();
-
-                for (coeff, label) in lc.iter() {
-                    let eval = match label {
-                        LCTerm::One => F::one(),
-                        LCTerm::PolyLabel(l) => *poly_evals
-                            .get(&(l.clone().into(), point.clone()))
-                            .ok_or(Error::MissingEvaluation { label: l.clone() })?,
-                    };
-
-                    actual_rhs += &(*coeff * &eval);
-                }
-                if claimed_rhs != actual_rhs {
-                    eprintln!("Claimed evaluation of {} is incorrect", lc.label());
-                    return Ok(false);
-                }
-            }
-        }
-
-        let pc_result = Self::batch_check_individual_opening_challenges(
-            vk,
-            commitments,
-            &poly_query_set,
-            &poly_evals,
-            proof,
-            opening_challenges,
-            rng,
-        )?;
-        if !pc_result {
-            eprintln!("Evaluation proofs failed to verify");
-            return Ok(false);
-        }
-
-        Ok(true)
-    }
+        Self::Commitment: 'a;
 }
 
 /// Evaluate the given polynomials at `query_set`.
